Extracted function to prepare for japanese
